refactor(cgt): consolidate duplicate helpers in calculation engine
Apply code-simplifier pattern to reduce code duplication:
- Extract groupBySymbol, getRemainingQuantity, calculateCostBasis to utils.ts
- Remove 4 duplicate implementations from matcher files (~150 lines)
- Maintain backward-compatible exports from sameDayMatcher and shortSellMatcher

All 339 unit tests and E2E tests pass.
